feat: Add spatie/laravel-sitemap package for generating sitemap.xml
This commit adds the spatie/laravel-sitemap package to the composer.json file, with a version constraint of "^7.2". This package is used for generating a sitemap.xml file for the website.

Additionally, a new route is added in the routes/web.php file to handle the "/sitemap.xml" URL. This route generates a sitemap using the spatie/laravel-sitemap package, including the homepage ("/"), the dashboard ("/beranda"), and the articles ("/articles"). It also dynamically adds URLs for each article in the database.

This commit enhances the website's SEO by providing a sitemap.xml file that search engines can use to discover and index the website's pages.