feat: Register SitemapServiceProvider in AppServiceProvider
This commit registers the `SitemapServiceProvider` in the `AppServiceProvider` class. By adding the line `$this->app->register(SitemapServiceProvider::class);` in the `register` method, the `SitemapServiceProvider` is now available for use in the application.

This change ensures that the `SitemapServiceProvider` is properly registered and can be used to generate the sitemap.xml file for the website.